Integrate GitHub changes
 DVS virtual: http://builds4u/dvs/#/change/2161036637926856.2?showTab=DVS
 bug 1865408

Jobs: 1865408-2006
[git-p4: depot-paths = "//sw/gpgpu/thrust/": change = 21611734]

+++ testing/reduce.cu
@@ -194,7 +194,7 @@ template<typename T>
   void TestReduceCountingIterator(size_t n)
 {
   // be careful not to generate a range larger than we can represent
-  n = thrust::min<size_t>(n, std::numeric_limits<T>::max());
+  n = thrust::min<size_t>(n, static_cast<size_t>(std::numeric_limits<T>::max()));
 
   thrust::counting_iterator<T, thrust::host_system_tag>   h_first = thrust::make_counting_iterator<T>(0);
   thrust::counting_iterator<T, thrust::device_system_tag> d_first = thrust::make_counting_iterator<T>(0);
