+++ testing/transform.cu
@@ -749,7 +749,7 @@ template <class T>
     KNOWN_FAILURE;
 #else
     // be careful not to generate a range larger than we can represent
-    n = thrust::min<size_t>(n, std::numeric_limits<T>::max());
+    n = thrust::min<size_t>(n, static_cast<size_t>(std::numeric_limits<T>::max()));
 
     thrust::counting_iterator<T, thrust::host_system_tag>   h_first = thrust::make_counting_iterator<T>(0);
     thrust::counting_iterator<T, thrust::device_system_tag> d_first = thrust::make_counting_iterator<T>(0);
@@ -774,7 +774,7 @@ template <typename T>
     KNOWN_FAILURE;
 #else
     // be careful not to generate a range larger than we can represent
-    n = thrust::min<size_t>(n, std::numeric_limits<T>::max());
+    n = thrust::min<size_t>(n, static_cast<size_t>(std::numeric_limits<T>::max()));
 
     thrust::counting_iterator<T, thrust::host_system_tag>   h_first = thrust::make_counting_iterator<T>(0);
     thrust::counting_iterator<T, thrust::device_system_tag> d_first = thrust::make_counting_iterator<T>(0);
